+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tool streaming support to `OllamaPromptDriver`.
 - `DateTimeTool.add_timedelta` and `DateTimeTool.get_datetime_diff` for basic datetime arithmetic.
 - Support for `pydantic.BaseModel`s anywhere `schema.Schema` is supported.
+- Support for `AudioArtifact` inputs/outputs in `OpenAiChatPromptDriver`.
 
 ### Changed